requirements.txt
babel==2.17.0
docxcompose==1.4.0
docxtpl==0.20.1
Jinja2==3.1.6
lxml==6.0.0
MarkupSafe==3.0.2
python-docx==1.2.0
setuptools==80.9.0
six==1.17.0
typing_extensions==4.14.1
openai==1.51.2
requests==2.32.3
python-dotenv==1.0.1
httpx==0.27.2
httpcore==1.0.5
